aider: Added hover text to the icon buttons for "delete highlight" and "generate flashcard from highlight."
# Aider chat conversation:

USER: Can you add some hover text to the icon buttons ("delete highlight" and "generate flashcard from highlight" respectively)

ASSISTANT: To add hover text to the icon buttons for "delete highlight" and "generate flashcard from highlight," I will add the `title` attribute to the `IconButton` components that wrap these icons in the `BookViewer.js` file.

Here are the diffs for the changes:

```diff
--- client/src/components/BookViewer.js
+++ client/src/components/BookViewer.js
@@ ... @@
 const renderHighlightPopover = () => (
   <Popover
     open={Boolean(anchorEl)}
     anchorEl={anchorEl}
     onClose={handlePopoverClose}
     anchorOrigin={{
       vertical: 'top',
       horizontal: 'center',
     }}
     transformOrigin={{
       vertical: 'bottom',
       horizontal: 'center',
     }}
   >
-    <Button
+    <Button
+      title="Delete Highlight"
       startIcon={<DeleteOutlineIcon />}
       onClick={() => {deleteHighlight(highlightToDelete._id); handlePopoverClose()}}
       color="error"
       size="small"
     >
     </Button>
-    <Button
+    <Button
+      title="Generate Flashcard"
       size="small"
       startIcon={<QuizIcon />}
       onClick={() => {generateFlashcards(); handlePopoverClose()}}
     >
     </Button>
   </Popover>
 );
```

This change adds a tooltip with the text "Delete Highlight" to the delete button and "Generate Flashcard" to the generate flashcard button.