aider: Updated the Sidebar component to make flashcards clickable and reuse the onHighlightClick method to jump to the corresponding highlight in the BookViewer.
# Aider chat conversation:

USER: Finally, reuse the `onHighlightClick` method as an onclick for the flascards (so clicking on a flashcard in the sidebar jumps to that highlight in the viewer).

ASSISTANT: To make the flashcards in the `Sidebar` component clickable and reuse the `onHighlightClick` method to jump to the corresponding highlight in the `BookViewer`, I will update the `Sidebar` component to include an `onClick` handler for each flashcard that calls `onHighlightClick` with the `highlightId` associated with the flashcard.

Here is the diff for the change:

```diff
--- client/src/components/Sidebar.js
+++ client/src/components/Sidebar.js
@@ ... @@
 function Sidebar({ highlights, onHighlightClick, flashcards }) {
   // ... existing code ...

   return (
     <div style={{ width: '250px', overflowY: 'auto' }}>
       {/* ... existing Tabs and TabPanel for highlights ... */}
       <TabPanel value={tabValue} index={1}>
         <List>
           {flashcards.map((flashcard) => (
             <React.Fragment key={flashcard._id}>
-              <ListItem alignItems="flex-start">
+              <ListItem alignItems="flex-start" button onClick={() => onHighlightClick(flashcard.highlightId)}>
                 <ListItemText
                   primary={flashcard.frontText}
                   secondary={flashcard.backText}
                 />
               </ListItem>
               <Divider />
             </React.Fragment>
           ))}
         </List>
       </TabPanel>
     </div>
   );
 }

 export default Sidebar;
```

This change will allow users to click on a flashcard in the sidebar to jump to the corresponding highlight in the `BookViewer`.